Added Favourites to Bckend Api

+++ pages/api/user/prompts/[id].js
@@ -4,6 +4,16 @@ import Product from "@/models/products";
 export default async function handler(req, res) {
     connectMongo();
     const { id } = req.query
+    const session = await getServerAuthSession(req, res)
+    if (!session) 
+    {
+        return res.status(401).json({ error: 'You are not authorized' })
+    }
+    const product= Product.findById(id)
+    if (!product.vendorId===session.user.id) {
+      return res.status(403).json({error:'Action is Forbidden'})
+      
+    }
     if (req.method==='DELETE'){ 
         Product.findByIdAndDelete(id)
       .then((p)=>{return res.status(200).json({message:p.name+" is Deleted Successfully"})})
